Move Steam script to a new platform directory
